% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CARME-package.R
\docType{package}
\name{CARME-package}
\alias{CARME-package}
\alias{CARME}
\title{The 'CARME' package.}
\description{
CAR-MM modelling in Stan
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.26.11. https://mc-stan.org
}
