% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_mm_high_ord.R
\name{f_mm_high_ord}
\alias{f_mm_high_ord}
\title{Helper function for MM matrix simulation}
\usage{
f_mm_high_ord(W, ord, mat_ord_weight, ind_range)
}
\description{
Helper function for MM matrix simulation
}
\keyword{internal}
