% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/W_sel.R
\docType{data}
\name{W_sel}
\alias{W_sel}
\title{Adjacency matrix for the South East London set of MSOAs}
\format{
A 152x152 symmetric matrix
}
\usage{
data(W_sel)
}
\description{
Adjacency matrix of 152 MSOAs in South East London, used for the data
analysis in the paper "Structure induced by a multiple membership
transformation on the Conditional Autoregressive model". Column and rows
names indicate the MSOA code.
}
\references{
Structure induced by a multiple membership transformation on
the Conditional Autoregressive model (2022)
(\href{https://arxiv.org/abs/2208.06738}{arXiv})
}
\keyword{datasets}
