% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_mm_high_ord.R
\name{f_mm_high_ord}
\alias{f_mm_high_ord}
\title{Helper function for MM matrix simulation}
\usage{
f_mm_high_ord(W, ord, mat_ord_weight, ind_range)
}
\description{
Helper function for MM matrix simulation
}
\references{
Marco Gramatica. Silvia Liverani. Peter Congdon.
"Structure Induced by a Multiple Membership Transformation on the Conditional
Autoregressive Model." Bayesian Analysis Advance Publication 1 - 25, 2023.
https://doi.org/10.1214/23-BA1370
}
\keyword{internal}
