% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_car.R
\name{sim_car}
\alias{sim_car}
\title{Simulation of proper CAR random effects}
\usage{
sim_car(W, alpha = 0.5, tau = 5)
}
\arguments{
\item{W}{Symmetric adjacency matrix of size \code{n}}

\item{alpha}{properness parameter between 0 and 1. Defaults to 0.5}

\item{tau}{marginal precision. Defaults to 5}
}
\value{
a vector of length \code{n}
}
\description{
\code{sim_car} returns a vector of CAR distributed random effects
}
\examples{
data(W_sel)
sim_car(W = W_sel, alpha = .9, tau = 5)
}
\references{
Jin, X., Carlin, B.P. and Banerjee, S. (2005), Generalized
Hierarchical Multivariate CAR Models for Areal Data. Biometrics,
61: 950-961. https://doi.org/10.1111/j.1541-0420.2005.00359.x
}
