% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/W_sel.R
\docType{data}
\name{W_sel}
\alias{W_sel}
\title{Adjacency matrix for the South East London set of MSOAs}
\format{
A 152x152 symmetric matrix
}
\usage{
data(W_sel)
}
\description{
Adjacency matrix of 152 MSOAs in South East London, used for the data
analysis in the paper "Structure induced by a multiple membership
transformation on the Conditional Autoregressive model". Column and rows
names indicate the MSOA code.
}
\references{
Marco Gramatica. Silvia Liverani. Peter Congdon.
"Structure Induced by a Multiple Membership Transformation on the Conditional
Autoregressive Model." Bayesian Analysis Advance Publication 1 - 25, 2023.
https://doi.org/10.1214/23-BA1370
}
\keyword{datasets}
